+++ libs/ts-rest/fastify/src/lib/ts-rest-fastify.ts
@@ -1,19 +1,15 @@
 import {
-  ApiRouteServerResponse,
   AppRoute,
   AppRouteMutation,
   AppRouteQuery,
   AppRouter,
   checkZodSchema,
-  LowercaseKeys,
   parseJsonQueryObject,
-  PathParamsWithCustomValidators,
+  ServerInferRequest,
+  ServerInferResponses,
   validateResponse,
-  Without,
-  ZodInferOrType,
 } from '@ts-rest/core';
 import * as fastify from 'fastify';
-
 import { z } from 'zod';
 
 export class RequestValidationError extends Error {
@@ -27,45 +23,12 @@ export class RequestValidationError extends Error {
   }
 }
 
-type AppRouteQueryImplementation<T extends AppRouteQuery> = (
-  input: Without<
-    {
-      params: PathParamsWithCustomValidators<T>;
-      query: ZodInferOrType<T['query']>;
-      headers: LowercaseKeys<ZodInferOrType<T['headers']>> &
-        fastify.FastifyRequest['headers'];
-      request: fastify.FastifyRequest;
-      reply: fastify.FastifyReply;
-    },
-    never
-  >
-) => Promise<ApiRouteServerResponse<T['responses']>>;
-
-type WithoutFileIfMultiPart<T extends AppRouteMutation> =
-  T['contentType'] extends 'multipart/form-data'
-    ? Without<ZodInferOrType<T['body']>, File>
-    : ZodInferOrType<T['body']>;
-
-type AppRouteMutationImplementation<T extends AppRouteMutation> = (
-  input: Without<
-    {
-      params: PathParamsWithCustomValidators<T>;
-      query: ZodInferOrType<T['query']>;
-      body: WithoutFileIfMultiPart<T>;
-      headers: LowercaseKeys<ZodInferOrType<T['headers']>> &
-        fastify.FastifyRequest['headers'];
-      request: fastify.FastifyRequest;
-      reply: fastify.FastifyReply;
-    },
-    never
-  >
-) => Promise<ApiRouteServerResponse<T['responses']>>;
-
-type AppRouteImplementation<T extends AppRoute> = T extends AppRouteMutation
-  ? AppRouteMutationImplementation<T>
-  : T extends AppRouteQuery
-  ? AppRouteQueryImplementation<T>
-  : never;
+type AppRouteImplementation<T extends AppRoute> = (
+  input: ServerInferRequest<T, fastify.FastifyRequest['headers']> & {
+    request: fastify.FastifyRequest;
+    reply: fastify.FastifyReply;
+  }
+) => Promise<ServerInferResponses<T>>;
 
 type RecursiveRouterObj<T extends AppRouter> = {
   [TKey in keyof T]: T[TKey] extends AppRouter
@@ -73,6 +36,11 @@ type RecursiveRouterObj<T extends AppRouter> = {
     : T[TKey] extends AppRoute
     ? AppRouteImplementation<T[TKey]>
     : never;
+};
+
+type InitialisedRouter<TContract extends AppRouter> = {
+  contract: TContract;
+  routes: RecursiveRouterObj<TContract>;
 };
 
 type RegisterRouterOptions = {
@@ -137,9 +105,15 @@ const validateRequest = (
 };
 
 export const initServer = () => ({
-  router: <T extends AppRouter>(router: T, args: RecursiveRouterObj<T>) => args,
+  router: <TContract extends AppRouter>(
+    contract: TContract,
+    routes: RecursiveRouterObj<TContract>
+  ): InitialisedRouter<TContract> => ({
+    contract,
+    routes,
+  }),
   registerRouter: <
-    T extends RecursiveRouterObj<TContract>,
+    T extends InitialisedRouter<TContract>,
     TContract extends AppRouter
   >(
     contract: TContract,
@@ -152,12 +126,34 @@ export const initServer = () => ({
       requestValidationErrorHandler: 'combined',
     }
   ) => {
-    recursivelyRegisterRouter(routerImpl, contract, [], app, options);
+    recursivelyRegisterRouter(routerImpl.routes, contract, [], app, options);
 
     app.setErrorHandler(
       requestValidationErrorHandler(options.requestValidationErrorHandler)
     );
   },
+  plugin:
+    <T extends AppRouter>(
+      router: InitialisedRouter<T>
+    ): fastify.FastifyPluginCallback<RegisterRouterOptions> =>
+    (
+      app,
+      opts = {
+        logInitialization: true,
+        jsonQuery: false,
+        responseValidation: false,
+        requestValidationErrorHandler: 'combined',
+      },
+      done
+    ) => {
+      recursivelyRegisterRouter(router.routes, router.contract, [], app, opts);
+
+      app.setErrorHandler(
+        requestValidationErrorHandler(opts.requestValidationErrorHandler)
+      );
+
+      done();
+    },
 });
 
 const requestValidationErrorHandler = (
@@ -183,14 +179,17 @@ const requestValidationErrorHandler = (
   };
 };
 
+type AppRouteMutationWithParams = AppRouteMutation & { path: '/:placeholder' };
+
 /**
  *
  * @param routeImpl - User's implementation of the route
  * @param appRoute - the `ts-rest` contract for this route (e.g. with Path, params, query, body, etc.)
  * @param fastify - the fastify instance to register the route on
+ * @param options - options for the routers
  */
 const registerRoute = <TAppRoute extends AppRoute>(
-  routeImpl: AppRouteImplementation<AppRoute>,
+  routeImpl: AppRouteImplementation<AppRouteMutationWithParams>,
   appRoute: TAppRoute,
   fastify: fastify.FastifyInstance,
   options: RegisterRouterOptions
@@ -217,7 +216,8 @@ const registerRoute = <TAppRoute extends AppRoute>(
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         headers: validationResults.headersResult.data as any,
         request,
-        body: request.body,
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        body: request.body as any,
         reply,
       });