+++ libs/ts-rest/nest/src/lib/ts-rest-nest.ts
@@ -2,17 +2,13 @@ import {
   AppRoute,
   AppRouter,
   Without,
-  ApiRouteServerResponse,
-  AppRouteStrictStatusCodes,
-  Extends,
+  ServerInferResponses,
 } from '@ts-rest/core';
 import { TsRestRequestShape } from './ts-rest-request.decorator';
 
 type AppRouterMethodShape<T extends AppRoute> = (
   ...args: any[]
-) => Promise<
-  ApiRouteServerResponse<T['responses'], Extends<T, AppRouteStrictStatusCodes>>
->;
+) => Promise<ServerInferResponses<T>>;
 
 type AppRouterControllerShape<T extends AppRouter> = Without<
   {
@@ -31,12 +27,7 @@ type AppRouterRequestShapes<T extends AppRouter> = Without<
 type AppRouterResponseShapes<T extends AppRouter> = Without<
   {
     [K in keyof T]: T[K] extends AppRoute
-      ? Promise<
-          ApiRouteServerResponse<
-            T['responses'],
-            Extends<T, AppRouteStrictStatusCodes>
-          >
-        >
+      ? ServerInferResponses<T[K]>
       : never;
   },
   never