+++ libs/ts-rest/nest/src/lib/ts-rest-request.decorator.ts
@@ -7,45 +7,32 @@ import {
   AppRoute,
   AppRouteMutation,
   checkZodSchema,
-  LowercaseKeys,
   parseJsonQueryObject,
-  PathParamsWithCustomValidators,
-  Without,
+  ServerInferRequest,
   zodErrorResponse,
-  ZodInferOrType,
 } from '@ts-rest/core';
 import type { Request } from 'express-serve-static-core';
 import { JsonQuerySymbol, TsRestAppRouteMetadataKey } from './constants';
 
-type BodyWithoutFileIfMultiPart<T extends AppRouteMutation> =
-  T['contentType'] extends 'multipart/form-data'
-    ? Without<ZodInferOrType<T['body']>, File>
-    : ZodInferOrType<T['body']>;
-
-export type TsRestRequestShape<TRoute extends AppRoute> = Without<
-  {
-    params: PathParamsWithCustomValidators<TRoute>;
-    body: TRoute extends AppRouteMutation
-      ? BodyWithoutFileIfMultiPart<TRoute>
-      : never;
-    query: ZodInferOrType<TRoute['query']>;
-    headers: LowercaseKeys<ZodInferOrType<TRoute['headers']>> &
-      Request['headers'];
-  },
-  never
+export type TsRestRequestShape<TRoute extends AppRoute> = ServerInferRequest<
+  TRoute,
+  Request['headers']
 >;
+
+type AppRouteMutationWithParams = AppRouteMutation & { path: '/:placeholder' };
 
 /**
  * Parameter decorator used to parse, validate and return the typed request object
  */
 export const TsRestRequest = createParamDecorator(
-  (_: unknown, ctx: ExecutionContext): TsRestRequestShape<any> => {
+  (
+    _: unknown,
+    ctx: ExecutionContext
+  ): TsRestRequestShape<AppRouteMutationWithParams> => {
     const req: Request = ctx.switchToHttp().getRequest();
 
-    const appRoute: AppRoute | undefined = Reflect.getMetadata(
-      TsRestAppRouteMetadataKey,
-      ctx.getHandler()
-    );
+    const appRoute: AppRouteMutationWithParams | undefined =
+      Reflect.getMetadata(TsRestAppRouteMetadataKey, ctx.getHandler());
 
     if (!appRoute) {
       // this will respond with a 500 error without revealing this error message in the response body
@@ -85,7 +72,7 @@ export const TsRestRequest = createParamDecorator(
 
     const bodyResult = checkZodSchema(
       req.body,
-      appRoute.method === 'GET' ? null : appRoute.body
+      (appRoute as AppRoute).method === 'GET' ? null : appRoute.body
     );
 
     if (!bodyResult.success) {
@@ -94,8 +81,8 @@ export const TsRestRequest = createParamDecorator(
 
     return {
       query: queryResult.data,
-      params: pathParamsResult.data,
-      body: bodyResult.data,
+      params: pathParamsResult.data as any,
+      body: bodyResult.data as any,
       headers: headersResult.data as TsRestRequestShape<
         typeof appRoute
       >['headers'],