+++ libs/ts-rest/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ts-rest/core
 
+## 3.25.1
+
+### Patch Changes
+
+- 81560d4: Fix ESM/CJS issues in package.json
+
+## 3.25.0
+
+### Patch Changes
+
+- bf21a75: Internal refactor of types
+
 ## 3.24.0
 
 ## 3.23.0