+++ libs/ts-rest/fastify/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ts-rest/fastify
 
+## 3.25.1
+
+### Patch Changes
+
+- 81560d4: Fix ESM/CJS issues in package.json
+
+## 3.25.0
+
+### Minor Changes
+
+- 871466c: Added support for registering ts-rest as a Fastify plugin
+
+### Patch Changes
+
+- bf21a75: Internal refactor of types
+
 ## 3.24.0
 
 ### Minor Changes